Added more features
Square Root Calculation
Factorial Calculation
Greatest Common Divisor (GCD)
Trigonometric Functions (sin, cos, tan)
User Input Support
